added edit component
