+++ server/index.js
@@ -93,7 +93,7 @@ app.get("/values/all", async (req, res) => {
   }
 });
 
-// Add a new endpoint to fetch unique roles
+
 app.get("/api/roles", async (req, res) => {
   try {
     const roles = await pgClient.query("SELECT DISTINCT UNNEST(role) AS role FROM values");
@@ -135,8 +135,6 @@ app.get("/values/:id", async (req, res) => {
 })
 
 
-
-// now the post -> insert value
 app.post("/values", async (req, res) => {
   const { account, sector, engagement, startdate, enddate, channel, owner, originator, role, location, revenue, forecast, notes, grade } = req.body
   if (!account || !sector || !engagement || !startdate || !enddate || !channel || !owner || !originator || !role || !location || !revenue || !forecast || !grade) {
